changed auth server to typescript
