--- auth-server/.eslintrc.js
@@ -1,23 +1,0 @@
-module.exports = {
-  env: {
-    es6: true,
-    node: true
-  },
-  extends: [
-    'standard'
-  ],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly'
-  },
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaVersion: 2018,
-    sourceType: 'module'
-  },
-  plugins: [
-    '@typescript-eslint'
-  ],
-  rules: {
-  }
-}
