+++ auth-server/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 .env
-data
+data/
+build/
