--- auth-server/db/schema.js
@@ -1,51 +1,0 @@
-// const { promisify } = require('util')
-const Knex = require('knex')
-const connection = require('../knexfile')
-const { Model } = require('objection')
-const bcrypt = require('bcrypt')
-// const crypto = require('crypto')
-
-const knexConnection = Knex(connection)
-
-Model.knex(knexConnection)
-
-class User extends Model {
-  static get tableName () {
-    return 'users'
-  }
-
-  static get idColumn () {
-    return 'id'
-  }
-
-  getUser () {
-    return {
-      email: this.email,
-      id: this.id,
-      username: this.username
-    }
-  }
-
-  async $beforeInsert () {
-    const salt = bcrypt.genSaltSync()
-    this.password = await bcrypt.hash(this.password, salt)
-  }
-
-  verifyPassword (password, callback) {
-    bcrypt.compare(password, this.password, callback)
-  }
-
-  static get jsonSchema () {
-    return {
-      type: 'object',
-      required: ['username'],
-      properties: {
-        email: { type: 'string', minLength: 1, maxLength: 255 },
-        id: { type: 'integer' },
-        username: { type: 'string', minLength: 1, maxLength: 255 }
-      }
-    }
-  }
-}
-
-module.exports = { User }
